CAMEL-4668: Fixed JPA consumer to rollback if one Exchange failed
git-svn-id: https://svn.apache.org/repos/asf/camel/branches/camel-2.7.x@1200880 13f79535-47bb-0310-9956-ffa450edef68

+++ components/camel-jpa/src/main/java/org/apache/camel/component/jpa/JpaConsumer.java
@@ -83,7 +83,11 @@ public class JpaConsumer extends ScheduledPollConsumer implements BatchConsumer,
 
                 Query query = getQueryFactory().createQuery(entityManager);
                 configureParameters(query);
+                LOG.trace("Created query {}", query);
+
                 List<Object> results = CastUtils.cast(query.getResultList());
+                LOG.trace("Got result list from query {}", results);
+
                 for (Object result : results) {
                     DataHolder holder = new DataHolder();
                     holder.manager = entityManager;
@@ -96,9 +100,14 @@ public class JpaConsumer extends ScheduledPollConsumer implements BatchConsumer,
                 try {
                     messagePolled = processBatch(CastUtils.cast(answer));
                 } catch (Exception e) {
-                    throw new PersistenceException(e);
+                    if (e instanceof PersistenceException) {
+                        throw (PersistenceException) e;
+                    } else {
+                        throw new PersistenceException(e);
+                    }
                 }
 
+                LOG.debug("Flushing EntityManager");
                 entityManager.flush();
                 return messagePolled;
             }
@@ -140,11 +149,12 @@ public class JpaConsumer extends ScheduledPollConsumer implements BatchConsumer,
                 if (LOG.isDebugEnabled()) {
                     LOG.debug("Processing exchange: " + exchange);
                 }
-                try {
-                    getProcessor().process(exchange);
-                } catch (Exception e) {
-                    throw new PersistenceException(e);
+                getProcessor().process(exchange);
+                if (exchange.getException() != null) {
+                    // if we failed then throw exception
+                    throw exchange.getException();
                 }
+
                 getDeleteHandler().deleteObject(entityManager, result);
             }
         }
